+++ tests/spec/directives/user-picker.spec.js
@@ -80,8 +80,8 @@ describe('luidUserPicker', function(){
 			isolateScope.find();
 			$httpBackend.flush();
 
-			users = [{"id":1,"firstName":"Guillaume","lastName":"Allain"},{"id":2,"firstName":"Elsa","lastName":"Arrou-Vignod"},{"id":3,"firstName":"Chloé","lastName":"Azibert Yekdah"},{"id":4,"firstName":"Clément","lastName":"Barbotin"}];
-			expect(angular.equals(isolateScope.users, users)).toBe(true);
+			usersIds = [1,2,3,4];
+			expect(angular.equals(_.pluck(isolateScope.users, 'id'), usersIds)).toBe(true);
 		});
 		it('should handle errors', function(){
 			spyOn(console,'log');
@@ -355,9 +355,9 @@ describe('luidUserPicker', function(){
 			$httpBackend.whenGET(/api\/v3\/users\?id=1,3\&fields=.*/i).respond(200, RESPONSE_2_homonyms_details_0_1);
 			$httpBackend.flush();
 
-			users = [{"id":1,"firstName":"Lucien","lastName":"Bertin","hasHomonyms":true,"mail":"no-reply@lucca.fr","employeeNumber":87,"legalEntity":{"name":"Lucca UK"},"department":{"name":"BU Timmi/Lucca"}},{"id":3,"firstName":"Lucien","lastName":"Bertin","hasHomonyms":true,"mail":"no-reply@lucca.fr","employeeNumber":110,"legalEntity":{"name":"Lucca"},"department":{"name":"Sales"}}];
 			var homonyms = _.where(isolateScope.users, {hasHomonyms:true}); // keep only the one having an homonym
-			expect(angular.equals(users, homonyms)).toBe(true);
+			expect(angular.equals(RESPONSE_2_homonyms_details_0_1.data.items[0].legalEntity, homonyms[0].legalEntity)).toBe(true);
+			// has added the LE to the first homonym
 		});
 		it('should handle errors when getting homonyms details', function(){
 			spyOn(console, 'log');
@@ -461,9 +461,9 @@ describe('luidUserPicker', function(){
 			$httpBackend.expectGET(/api\/v3\/users\?id=1,3\&fields=id,firstname,lastname,birthDate,mail,manager.name/i).respond(RESPONSE_2_homonyms_details_0_2);
 			$httpBackend.flush();
 
-			users = [{"id":1,"firstName":"Lucien","lastName":"Bertin","hasHomonyms":true,"mail":"no-reply@lucca.fr","manager":{"name":"Romain Vergnory"},"birthDate":"1990-12-10T00:00:00"},{"id":3,"firstName":"Lucien","lastName":"Bertin","hasHomonyms":true,"mail":"no-reply@lucca.fr","manager":{"name":"Benoît Paugam"},"birthDate":"1986-03-25T00:00:00"}];
+			// users = [{"id":1,"firstName":"Lucien","lastName":"Bertin","hasHomonyms":true,"mail":"no-reply@lucca.fr","manager":{"name":"Romain Vergnory"},"birthDate":"1990-12-10T00:00:00"},{"id":3,"firstName":"Lucien","lastName":"Bertin","hasHomonyms":true,"mail":"no-reply@lucca.fr","manager":{"name":"Benoît Paugam"},"birthDate":"1986-03-25T00:00:00"}];
 			var homonyms = _.where(isolateScope.users, {hasHomonyms:true}); // keep only the one having an homonym
-			expect(angular.equals(users, homonyms)).toBe(true);
+			expect(angular.equals(RESPONSE_2_homonyms_details_0_2.data.items[0].birthDate, homonyms[0].birthDate)).toBe(true);
 		});
 		it('should identify the first and third property as differentiating properties', function(){
 			$httpBackend.expectGET(/api\/v3\/users\?id=1,3\&fields=id,firstname,lastname,birthDate,mail,manager.name/i).respond(RESPONSE_2_homonyms_details_0_2);
@@ -568,6 +568,59 @@ describe('luidUserPicker', function(){
 				.first(5) // we exclude the overflow message
 				.value();
 			expect(customInfos).toEqual([2,4,6,8,10]);
+		});
+	});
+
+	/********************************
+	** DISPLAY SELECTED USER FIRST **
+	********************************/
+	describe("with selected user", function(){
+		var chloe = { id:3,
+			firstName:"Chloé",
+			lastName:"Azibert Yekdah"
+		};
+		var sandrine = { id:11,
+			firstName:"Sandrine",
+			lastName:"Conraux"
+		};
+		beforeEach(function(){
+			$scope.myUser = chloe;
+			var tpl = angular.element('<luid-user-picker ng-model="myUser"></luid-user-picker>');
+			elt = $compile(tpl)($scope);
+			isolateScope = elt.isolateScope();
+			$scope.$digest();
+			isolateScope.find();
+			$httpBackend.whenGET(findApi).respond(200, RESPONSE_20_users);
+			$httpBackend.flush();
+		});
+		it("should call reorderUsers() when myUser is updated", function(){
+			spyOn(isolateScope, 'reorderUsers');
+			$scope.myUser = sandrine;
+			$scope.$digest();
+
+			expect(isolateScope.reorderUsers).toHaveBeenCalled();
+		});
+		it("should flag if the list of users returned by find contains the current myUser", function(){
+			expect(_.where(isolateScope.users, {isSelected:true}).length).toBe(1);
+			expect(_.first(isolateScope.users).isSelected).toBe(true);
+			expect(_.where(isolateScope.users, {isSelected:true})[0].id).toBe(chloe.id);
+		});
+		it("should have the right order of displayed users", function(){
+			var userIds = _.pluck(isolateScope.users, 'id');
+			expect(userIds).toEqual([3,1,2,4,5,-1]); // the -1 is because of the overflow
+		});
+		it("should update the selected one ", function(){
+			$scope.myUser = sandrine;
+			$scope.$digest();
+			expect(_.where(isolateScope.users, {isSelected:true}).length).toBe(1);
+			expect(_.first(isolateScope.users).isSelected).toBe(true);
+			expect(_.where(isolateScope.users, {isSelected:true})[0].id).toBe(sandrine.id);
+		});
+		it("should update the order of users ", function(){
+			$scope.myUser = sandrine;
+			$scope.$digest();
+			var userIds = _.pluck(isolateScope.users, 'id');
+			expect(userIds).toEqual([11,1,2,3,4,-1]); // the -1 is because of the overflow
 		});
 	});